factor out the output marshalling code.
